Completion of Chapter 9
Great chapter. Learnt how to store information in a class as well as how to create and import modules.
Now, obviously, more things were learnt in this chapter. It's just been the most tedious, possibly beacause we've been handed university coursework. 

I'll continue with the daily uploads of the exercises, however, it'll most likely be at the most 5 files daily. My time now is rather limited.

All the best.