Completion of Chapter 10
What a chapter!!! It's been the best one so far.

In this chapter we started of learning about and implementing code to handle errors(Exception Handling). Throughout to the end we continued to do so, and used what had been taught in the previous chapters to build rather impressive programs.

I must admit: I am rather excited for the things I'll create.

I had said I was only going to program for one hour, almost three hours ago.

All the best.